+++ src/main/java/com/iem/FilmRentalStore/entity/Actor.java
@@ -17,7 +17,7 @@ public class Actor {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "actor_id")
-    private Integer actorId;
+    private Short actorId;
 
     @Column(name = "first_name", nullable = false, length = 45)
     private String firstName;
